+++ arms-20190808/core/Models/CreateRumAppShrinkRequest.cs
@@ -46,6 +46,10 @@ namespace AlibabaCloud.SDK.ARMS20190808.Models
         [Validation(Required=false)]
         public string PackageName { get; set; }
 
+        [NameInMap("RealRegionId")]
+        [Validation(Required=false)]
+        public string RealRegionId { get; set; }
+
         /// <summary>
         /// The region ID.
         ///